[UPD] restructure of go files && [UPD] errors interfaces are now in use && [ADD] logrus && [FIX] minor constants issue
